refactor: extract Lexer class
This allows me to separate the generic functionality for building lexers
from the domain-specific rules that we need in order to parse JSP. The
interesting stuff is therefore all concentrated in `lex.js`, with the
uninteresting plumbing off in `Lexer.js`. For now, not even testing the
`Lexer` class, because it is invoked and tested heavily in the `lex.js`
tests.
