+++ packages/liferay-npm-scripts/test/utils/permute.js
@@ -1,3 +1,9 @@
+/**
+ * © 2019 Liferay, Inc. <https://liferay.com>
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
 const permute = require('../../src/utils/permute');
 
 describe('permute()', () => {
